requirements.txt
streamlit
python-dotenv
langgraph
langgraph-checkpoint-sqlite
langchain
langchain-core
langchain-community
langchain-google-genai
requests
ddgs